Addressing code review changes for PR #732:
* Replaced @exception decorator with try/except in main()
* Replaced optional --name with positional name and removed default
* Added license text to files
* Added django units to cover future API changes
* Refactored into submodules to better support tests
