+++ utils/cli/core/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: MIT
+from .definition import parser, ResourceType  # noqa
+from .core import CLI, CVAT_API_V1  # noqa
